[Experimental] Wire Quest sparse attention into the FlashAttention decode path
Adds a default-off --enable-sparse-attention flag that wires the existing
(previously dormant) sparse-algorithm framework (SparseCoordinator +
QuestAlgorithm + FlashAttentionAdaptor) into the non-MLA FlashAttention decode
path, so query-aware KV-cache page selection runs end to end.

- server_args: --enable-sparse-attention / --sparse-attention-config (enum via
  the existing _ALGORITHM_REGISTRY; decoupled from --enable-hisparse).
- factory: parse_sparse_attention_config(); create_sparse_coordinator() accepts a
  pre-parsed SparseConfig.
- model_runner: construct a SparseCoordinator for non-MLA models when enabled.
- scheduler / batch_result_processor: on_request_begin / on_request_end lifecycle.
- flashattention_backend: select top-k pages (attention_begin) + update page reps
  (attention_end) on decode; build prefill reps on extend.
- SparseConfig.min_sparse_prompt_len defaults to 0 (None crashed _compute_sparse_mask).
- test: CPU unit tests for config parsing + algorithm construction.

When the flag is off there is no behavior change. Decode CUDA graph is not yet
supported (run with --cuda-graph-backend-decode disabled); FA3 + non-MLA only.

+++ python/sglang/srt/mem_cache/sparsity/algorithms/base_algorithm.py
@@ -222,7 +222,11 @@ class BaseSparseAlgorithmImpl(BaseSparseAlgorithm):
         if layer_id == self.end_layer - 1:
             success_indices = req_pool_indices[valid_mask]
             self.states.repr_constructed[success_indices] = True
-            self.states.last_constructed_page[success_indices] = num_pages[valid_mask]
+            # seq_lens (hence num_pages) may be int32 while last_constructed_page
+            # is int64; cast to avoid an index_put dtype-mismatch error.
+            self.states.last_constructed_page[success_indices] = num_pages[
+                valid_mask
+            ].to(self.states.last_constructed_page.dtype)
 
     def update_representations(
         self,
@@ -257,7 +261,9 @@ class BaseSparseAlgorithmImpl(BaseSparseAlgorithm):
         # Update tracking states
         if layer_id == self.end_layer - 1:
             success_indices = req_pool_indices[valid_mask]
-            self.states.last_constructed_page[success_indices] = end_page[valid_mask]
+            self.states.last_constructed_page[success_indices] = end_page[
+                valid_mask
+            ].to(self.states.last_constructed_page.dtype)
 
     def retrieve_topk(
         self,
